requirements.txt
Flask
flask-cors
pymongo
requests
python-dotenv
Werkzeug